modifiche crud gestione menu

+++ src/main/java/it/decimo/merchant_service/controller/MenuController.java
@@ -10,8 +10,11 @@ import it.decimo.merchant_service.model.MenuItem;
 import it.decimo.merchant_service.service.MenuService;
 import it.decimo.merchant_service.service.MerchantService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import javax.websocket.server.PathParam;
 
 @RestController
 @RequestMapping("/api/merchant/{id}/menu")
@@ -21,6 +24,7 @@ public class MenuController {
     private MenuService menuService;
     @Autowired
     private MerchantService merchantService;
+
 
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Ritorna la lista di elementi che compongono il menu del locale", content = @Content(array = @ArraySchema(minItems = 0, uniqueItems = true, schema = @Schema(implementation = MenuItem.class)))),
@@ -35,13 +39,17 @@ public class MenuController {
 
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Ritorna l'id dell'oggetto che è stato inserito", content = @Content(schema = @Schema(implementation = BasicResponse.class))),
-            @ApiResponse(responseCode = "404", description = "Il ristorante ricercato non esite", content = @Content(schema = @Schema(implementation = BasicResponse.class)))})
+            @ApiResponse(responseCode = "404", description = "Il ristorante ricercato non esite", content = @Content(schema = @Schema(implementation = BasicResponse.class))),
+            @ApiResponse(responseCode = "401", description = "L'utente non può effettuare la richiesta perché non è il proprietario del merchant", content = @Content(schema = @Schema(implementation = BasicResponse.class)))})
     @PostMapping
-    public ResponseEntity<Object> insertItem(@PathVariable int id, @RequestBody MenuItem item) {
+    public ResponseEntity<Object> insertItem(@PathVariable int id, @RequestBody MenuItem item, @PathParam(value = "requester") int requester) {
         if (!merchantService.merchantExists(id)) {
             return ResponseEntity.status(404).body(new BasicResponse("No merchant found", "NO_MERCH_FOUND"));
         }
-
+        final var merchant = merchantService.getMerchant(id);
+        if (merchant.getOwner() != requester) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         final var saved = menuService.save(id, item);
 
         return ResponseEntity.ok().body(new BasicResponse(Integer.toString(saved.getMenuItemId()), "OK"));
@@ -49,13 +57,19 @@ public class MenuController {
 
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Ha rimosso l'oggetto dal menu del locale", content = @Content(schema = @Schema(implementation = BasicResponse.class))),
-            @ApiResponse(responseCode = "404", description = "Il locale non è stato trovato", content = @Content(schema = @Schema(implementation = BasicResponse.class)))})
+            @ApiResponse(responseCode = "404", description = "Il locale non è stato trovato", content = @Content(schema = @Schema(implementation = BasicResponse.class))),
+            @ApiResponse(responseCode = "401", description = "L'utente non può effettuare la richiesta perché non è il proprietario del merchant", content = @Content(schema = @Schema(implementation = BasicResponse.class)))})
 
     @DeleteMapping("/{itemId}")
     public ResponseEntity<Object> deleteMenuItem(@PathVariable(name = "id") int merchantId,
-                                                 @PathVariable(name = "itemId") int itemId) {
+                                                 @PathVariable(name = "itemId") int itemId, @PathParam(value = "requester") int requester) {
         if (!merchantService.merchantExists(merchantId)) {
             return ResponseEntity.status(404).body(new BasicResponse("No merchant found", "NO_MERCH_FOUND"));
+        }
+
+        final var merchant = merchantService.getMerchant(merchantId);
+        if (merchant.getOwner() != requester) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
 
         menuService.deleteItem(itemId, merchantId);
